Add UnitType to Inventory DB.
Remove supplier and category from the DB and add unitType , double quantity, barcode in the DB and business logic. ( In Item class and application Layer)
UI update -> add barcode search bar. (filter by filterByBarcode() )

+++ bill_item.h
@@ -13,7 +13,7 @@ class BillItem {
 public:
 
     // Constructor
-    explicit BillItem(int id, QString Name, int qty, double price)
+    explicit BillItem(int id, QString Name, double qty, double price)
         : m_ID(id), m_name(Name), m_quantity(qty), m_price_per_unit(price) {
 
         m_total_price = m_quantity * m_price_per_unit; // calculate total price of one Item when it added to bill
@@ -22,12 +22,12 @@ public:
     // Getters
     int getID() const { return m_ID; }
     QString getName() const { return m_name; }
-    int getQuantity() const { return m_quantity; }
+    double getQuantity() const { return m_quantity; }
     double getPricePerUnit() const { return m_price_per_unit; }
     double getTotalPrice() const { return m_total_price; }
 
     // Setter
-    void setQuantity(int qty) {
+    void setQuantity(double qty) {
         m_quantity = qty;
         m_total_price = m_quantity * m_price_per_unit; // Update total price when quantity changes
     }
@@ -35,7 +35,7 @@ public:
 private:
     int m_ID;
     QString m_name;
-    int m_quantity;
+    double m_quantity;
     double m_price_per_unit;
     double m_total_price; // total price of an item (same item)
